Register for AnyDown press as well
This is needed because addons no longer work on AnyUp any more unless
the cvar ActionButtonUseKeyDown is changed to false, forcing all actions
(including from addons) to work in the same key stroke.

Blizzard doesn't know how to fix this, and have applied dumb workarounds
for us so empowered spell casts and hold-to-cast works.

https://github.com/BigWigsMods/WoWUI/commit/198e2e1024fe45235882fe08597557576bc9954d#diff-076dae7c289d5bad160d1a99d98f56c41184deb56dbc286b6aa681ea1e1a926e